Correciones en ofertas

+++ app/src/main/java/com/example/culturaverde/Controllers/OfertasControlador.kt
@@ -11,7 +11,7 @@ interface OfertasControlador {
     @Headers("Content-Type: application/json", "Accept: application/json")
     @PUT("guardarOferta")
     fun guardarOferta(@Query("id_producto_productor") id_producto_productor:Long, @Query("porcentaje") porcentaje:Int,
-                      @Query("activo") activo:Boolean, @Query("id_oferta") id_oferta:Int ): Call<Void>
+                      @Query("activo") activo:Boolean, @Query("id_oferta") id_oferta: Long): Call<Void>
